Softened student and prepayment editors
